Updated versions of the PrimeCounter problem:
BasicPrimeCounter is what used to be PrimeCounter, and BestPrimeCounter is what used to be BetterPrimeCounter.  BetterPrimeCounter is a version of BasicPrimeCounter that only checks factors up to the square root of n as opposed to going all the way up to n, thus making it more efficient.

PrimeCounterTimeCheck allows us to compare the numbers obtained by each of the three PrimeCounter methods to confirm that they all yield the same values and thus all function correctly, as well as to compare the amounts of time each one takes.